PolicyManager Condensing
There were three different scripts where policy-related code was copied and pasted. This caused a lot of confusion as to where the game was actually using it. 
Now most of this is condensed into just the PolicyManager; some more needs to be condensed, but that will happen at a later time. 
This may need to happen with other things as well, but that is to be determined.